TAB-6385: Move maintenance mode calls into EntityRef
Moved the normal uses of EntityMaintenanceRef inside EntityRef and exposed the use of the ref (including its write-lock) as entity create/destroy methods.
This removes the ability for user code to directly acquire the EntityMaintenanceRef, relegating it to a purely internal utility class.  In a future change, it will be entirely removed in favor of more direct access to its underlying capabilities.