Issue 227: HA/Message Tracking changes
There are two things to track/inform. First, the client/transaction id
unique information needs to be communicated on each invoke to the passive,
so that things can be tracked as they are applied. This allows for
messages to be ignored on reconnect if they have already been seen.
Additionally, the eldest still active transaction id for the same
client is passed in, to allow for stepwise garbage collection.

Coarse grained GC is established by notifying passive clients of
client disconnect; at that point, all of the pending information
related to the at client can be forgotten.

invoke() was split into invokeActive() and invokePassive(), since
they have the same signature except for return type.